Squash 4 commits from fix/external-insurance-data-error that were added again after being removed due to a reverted PR:
* Create StartDateLabelSwitcher.tsx and move everything that has to do with that from StartDate.tsx. Then conditionally render StartDateLabelSwitcher only if there is a 'dataCollectionId' property on the 'quote'. This way there won't be an error from doing the externalInsuranceData query, since it will now only be called if we actually have the variable that we need to pass (the dataCollectionId).

* Add TODO comment in ExternalInsuranceData.graphql:
The fields we're querying for now ('dataCollection' and 'monthlyPremium') are deprecated and will not work when we implement Insurely in Norway. But it turned out to be not as easy as just replacing the fields but we'll probably have to write entirely new queries. And it's not reasonable to do that now.

* Update comment in StartDate - make it a TODO and make it clearer.

* Change condition for rendering <StartDateLabelSwitcher> in StartDate.tsx. From 'quote' having the 'dataCollectionId' property to it having 'currentInsurer', since the 'dataCollectionId' is null if the current insurer is not integrated with Insurely.
