+++ src/client/graphql/ExternalInsuranceData.graphql
@@ -1,12 +1,13 @@
+# TODO Replace 'dataCollection' and 'monthlyPremium' since they are deprecated and will not work when we implement Insurely in Norway
 query ExternalInsuranceData($reference: ID!) {
-    externalInsuranceProvider {
-        dataCollection(reference: $reference) {
-            renewalDate
-            insuranceProvider
-            monthlyPremium {
-                currency
-                amount
-            }
-        }
+  externalInsuranceProvider {
+    dataCollection(reference: $reference) {
+      renewalDate
+      insuranceProvider
+      monthlyPremium {
+        currency
+        amount
+      }
     }
+  }
 }
