+++ src/client/pages/OfferNew/Introduction/Sidebar/StartDate.tsx
@@ -1,7 +1,6 @@
 import { css } from '@emotion/core'
 import styled from '@emotion/styled'
 import { colorsV3 } from '@hedviginsurance/brand'
-import { externalInsuranceProviders } from '@hedviginsurance/embark'
 import { format, isToday, Locale, parse } from 'date-fns'
 import { motion } from 'framer-motion'
 import hexToRgba from 'hex-to-rgba'
@@ -17,11 +16,7 @@ import {
   useCurrentLocale,
   useMarket,
 } from 'components/utils/CurrentLocale'
-import {
-  useExternalInsuranceDataQuery,
-  useRemoveStartDateMutation,
-  useStartDateMutation,
-} from 'data/graphql'
+import { useRemoveStartDateMutation, useStartDateMutation } from 'data/graphql'
 import { CancellationOptions } from 'pages/OfferNew/Introduction/Sidebar/CancellationOptions'
 import { OfferData, OfferQuote } from 'pages/OfferNew/types'
 import {
@@ -32,6 +27,7 @@ import {
 import { useTextKeys } from 'utils/textKeys'
 import { Size } from 'components/types'
 import { gqlDateFormat } from './utils'
+import { StartDateLabelSwitcher } from './StartDateLabelSwitcher'
 
 interface Props {
   offerData: OfferData
@@ -100,16 +96,9 @@ const StartDateRowLabel = styled.div`
   padding-bottom: 0.5rem;
 `
 
-const DateLabel = styled.span<{ hasStartDate: boolean }>`
+const DateLabel = styled.span`
   margin-right: 5px;
   text-align: left;
-  ${({ hasStartDate }) =>
-    !hasStartDate &&
-    css`
-      font-size: 0.875rem;
-      line-height: 1;
-      color: ${colorsV3.gray500};
-    `};
 `
 
 const Value = styled.div`
@@ -129,24 +118,6 @@ const ErrorMessage = styled(motion.div)`
   padding: 1.25rem;
   margin-bottom: 10px;
   color: ${colorsV3.gray900};
-`
-
-const DataCollectedStartDateWrapper = styled.div`
-  display: flex;
-  align-items: flex-end;
-  flex-direction: column;
-  text-align: right;
-  margin-right: 0.5rem;
-`
-
-const DataCollectedStartDateDescription = styled.span`
-  font-size: 0.75rem;
-  line-height: 0.75rem;
-  color: ${colorsV3.gray700};
-`
-
-const DataCollectedStartDateValue = styled.span`
-  font-size: 1rem;
 `
 
 const DateInputModalWrapper = styled.div<{ isOpen: boolean }>`
@@ -203,11 +174,6 @@ const DateForm: React.FC<{
   const textKeys = useTextKeys()
   const [setStartDate] = useStartDateMutation()
   const [removeStartDate] = useRemoveStartDateMutation()
-  const { data: externalInsuranceData } = useExternalInsuranceDataQuery({
-    variables: {
-      reference: quote.dataCollectionId || '',
-    },
-  })
 
   const locale = useCurrentLocale()
   const market = useMarket()
@@ -237,35 +203,9 @@ const DateForm: React.FC<{
 
       return format(dateValue, getDateFormat(market)!, { locale: dateLocale! })
     }
-
-    const firstExternalInsurance =
-      externalInsuranceData?.externalInsuranceProvider?.dataCollection[0]
-    const renewalDate = firstExternalInsurance?.renewalDate
-
-    if (renewalDate) {
-      const externalInsuranceProvider = externalInsuranceProviders.find(
-        (provider: { externalCollectionId?: string }) =>
-          provider.externalCollectionId ===
-          firstExternalInsurance?.insuranceProvider?.toUpperCase(),
-      )
-
-      return (
-        <DataCollectedStartDateWrapper>
-          <DataCollectedStartDateValue>
-            {renewalDate}
-          </DataCollectedStartDateValue>
-          <DataCollectedStartDateDescription>
-            {textKeys.START_DATE_EXTERNAL_PROVIDER_SWITCH({
-              insuranceProvider: externalInsuranceProvider?.name,
-            })}
-          </DataCollectedStartDateDescription>
-        </DataCollectedStartDateWrapper>
-      )
-    }
-
-    return textKeys.SIDEBAR_STARTDATE_CELL_VALUE_SWITCHER()
   }
 
+  // TODO: make this function (which returns the calendar component) a regular React function component, in a file of its own
   const getDateInput = () => (
     <StyledDateInput
       open={datePickerOpen}
@@ -299,6 +239,7 @@ const DateForm: React.FC<{
   )
 
   const hasStartDate = Boolean(getDefaultDateValue(quote))
+
   return (
     <RowButtonWrapper>
       {isSplit && (
@@ -313,10 +254,15 @@ const DateForm: React.FC<{
         size={size}
       >
         <Value>
-          <DateLabel hasStartDate={Boolean(dateValue)}>
-            {getDateLabel()}
+          <DateLabel>
+            {!hasStartDate && hasCurrentInsurer(quote) && (
+              <StartDateLabelSwitcher
+                dataCollectionId={quote.dataCollectionId}
+              />
+            )}
+            {hasStartDate && getDateLabel()}
           </DateLabel>
-          {hasStartDate && <DownArrow />}
+          <DownArrow />
         </Value>
       </RowButton>
       {modal ? (
